+++ lsp/src/test/kotlin/org/gradle/declarative/lsp/VersionedDocumentStoreTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 the original author or authors.
+ * Copyright 2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package org.gradle.declarative.lsp.service
+package org.gradle.declarative.lsp
 
 import io.mockk.mockk
 import org.gradle.declarative.dsl.schema.AnalysisSchema
-import org.gradle.declarative.lsp.VersionedDocumentStore
 import org.gradle.internal.declarativedsl.dom.operations.overlay.DocumentOverlayResult
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
